Disable tab tracking while private browsing
Added setting “Track closed tabs while Private Browsing is on”.
Defaults to true so it doesn’t change current extension behavior.

Safari extension API doesn’t expose private browsing status at the
window or tab level. It’s all-or-nothing for the entire browser (as far
as the API is concerned). Would be better if I could stop tracking only
private windows, but not currently possible (unless I inject a bunch of
Javascript into each page that tests by attempting to write data to the
browser’s local storage, which would fail if private browsing is on for
that tab.)